Engenharia de Dados/ApiFundsExplorer/scraping/requirements.txt
beautifulsoup4
lxml
selenium>=4.18.0
pandas>=2.1.0
google-cloud-storage>=2.14.0